+++ NEWS.md
@@ -1,3 +1,11 @@
+# usincometaxes 0.7.0
+
+Update `taxsim.wasm` and `taxsim.js` to latest files. ETag of current version is `e23da-601f65ac97a40`.
+
+Add the following old dependent columns back to input columns: 'dep13', 'dep17', 'dep18'
+
+Add the following new columns to the output: 'credits' and 'staxbc'
+
 # usincometaxes 0.6.0
 
 Removed the `interface` option to `taxsim_calculate_taxes()`. Now, 'wasm' is the only interface option. Users are not able to send the data to the TAXSIM server via ssh or http. This feature was removed because we have seen unexpected changes to the TAXSIM server's output, which could silently introduce errors.